Organize Correspondence domain checks
