+++ K6/api/domain-checks/correspondence/correspondence.js
@@ -3,7 +3,7 @@ import { check } from "k6";
 import {
     CorrespondenceOverviewExt,
     InitializeCorrespondencesResponseExt,
-} from "../../clients/correspondence/correspondence.types.js";
+} from "../../../clients/correspondence/correspondence.types.js";
 
 function NormalizeRecipient(recipient) {
     if (/^\d{11}$/.test(recipient)) {
@@ -63,39 +63,6 @@ function CheckInitializedCorrespondences(response, expectedRecipients) {
         );
         console.error(
             `CheckInitializedCorrespondences - returned correspondences: ${JSON.stringify(correspondences)}`,
-        );
-    }
-
-    return success;
-}
-
-/**
- * Checks attachment ids returned by a correspondence upload operation.
- *
- * @param {InitializeCorrespondencesResponseExt} response API response.
- * @param {number} expectedCount Number of uploaded attachments.
- * @returns {boolean} True if the expected attachment ids were returned.
- */
-function CheckAttachmentIds(response, expectedCount) {
-    const attachmentIds = Array.isArray(response?.attachmentIds)
-        ? response.attachmentIds
-        : [];
-
-    const success = check(response, {
-        "CheckAttachmentIds - Expected number of attachment ids is returned":
-            () => attachmentIds.length === expectedCount,
-        "CheckAttachmentIds - Every attachment has an id": () =>
-            attachmentIds.length > 0 &&
-            attachmentIds.every(
-                (attachmentId) =>
-                    typeof attachmentId === "string" &&
-                    attachmentId.length > 0,
-            ),
-    });
-
-    if (!success) {
-        console.error(
-            `CheckAttachmentIds - expected ${expectedCount} attachment id(s), got: ${JSON.stringify(attachmentIds)}`,
         );
     }
 
@@ -245,7 +212,6 @@ function CheckMessageBody(messageBody) {
 
 export const CorrespondenceDomainChecks = {
     CheckInitializedCorrespondences,
-    CheckAttachmentIds,
     CheckCorrespondenceIds,
     CheckCorrespondenceOverview,
     FindDialogId,
